requirements_test.txt
coverage==4.5.3
pytest==4.6.3
pytest-django==3.5.0
pytest-cov==2.7.1
freezegun==0.3.12
